### 2023-05-18 #591 @yjunechoe
Presubmission Inquiry: jlmerclusterperm
Submitting Author Name: June Choe
Submitting Author Github Handle: @yjunechoe
Repository:  https://github.com/yjunechoe/jlmerclusterperm
Submission type: Pre-submission

---

```
Package: jlmerclusterperm
Title: Cluster-Based Permutation Analysis for Densely Sampled Time Data
Version: 1.0.0
    person("June", "Choe", , "jchoe001@gmail.com", role = c("aut", "cre", "cph"),
           comment = c(ORCID = "0000-0002-0701-921X"))
Description: An implementation of fast cluster-based permutation analysis
    (CPA) for densely-sampled time data developed in Maris & Oostenveld,
    2007 <doi:10.1016/j.jneumeth.2007.03.024>. Supports (generalized,
    mixed-effects) regression models for the calculation of timewise
    statistics. Provides both a wholesale and a piecemeal interface to the
    CPA procedure with an emphasis on interpretability and diagnostics.
    Integrates 'Julia' libraries 'MixedModels.JL' and 'GLM.JL' for
    performance improvements, with additional functionalities for
    interfacing with 'Julia' from 'R' powered by the 'JuliaConnectoR'
    package.
    https://yjunechoe.github.io/jlmerclusterperm/
    cli,
    generics,
    JuliaConnectoR,
    lme4,
    parallel,
    utils
    broom,
    broom.mixed,
    covr,
    eyetrackingR,
    forcats,
    future,
    MASS,
    patchwork,
    scales,
    tibble
    knitr
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
    "srr::srr_stats_roclet"))
SystemRequirements: Julia (>= 1.8)
Collate:
    'jlmerclusterperm-package.R'
    'aaa.R'
    'utils.R'
    'interop-utils.R'
    'interop-utils-unexported.R'
    'julia_rng.R'
    'jlmer_spec.R'
    'jlmer.R'
    'compute_timewise_statistics.R'
    'permute.R'
    'permute_timewise_statistics.R'
    'clusters_methods.R'
    'extract_clusters.R'
    'calculate_pvalue.R'
    'clusterpermute.R'
    'threshold_search.R'
    'tidy.R'
    'zzz.R'
    'srr-stats-standards.R'
```

jlmerclusterperm implements a regression-based CPA by fitting a regression model to each time point in a timeseries, and using the array of t-statistics from the regression output as a measure of "distance" between groups/lines in a time series. However, the package itself does not implement any novel regression algorithms - it simply calls [GLM.jl](https://github.com/JuliaStats/GLM.jl) (R equivalent lm/glm) and [MixedModels.jl](https://github.com/JuliaStats/MixedModels.jl) (R equivalent lme4::lmer/glmer) model fitting functions via the Julia backend.

Yes - I have started with a first draft. Tests specific to the correctness of the regression implementation itself are generally not applicable, to the extent that they are already tested internally in Julia by GLM and MixedModels themselves. However, correctness tests are included where the application of regression has consequences for the CPA procedure specifically.

Researchers in psychology, neuroscience, and linguistics conducting experiments on human subjects, working with densely-sampled time data (e.g., eye-tracking and EEG) with multi-level grouping structures (multiple by-participant and by-item observations). CPA was originally designed for application in EEG and have since been adopted for other methodologies such as eye-tracking and mouse-tracking.

Yes - [eyetrackingR](https://github.com/jwdink/eyetrackingR), [clusterperm](https://github.com/dalejbarr/clusterperm/), [permuco](https://jaromilfrossard.github.io/permuco/), and [permutes](https://github.com/cvoeten/permutes). eyetrackingR and clusterperm are inactive and no longer maintained as of late. permuco is feature-rich and maintained but does not support mixed models. permutes is a fast implementation of CPA and does support mixed models, but via an approximation algorithm which is not faithful to the original procedure as outlined in Maris & Oostenveld (2007). jlmerclusterperm is unique in supporting a mixed-models-based CPA that is computationally feasible to run and much faster than prior implementations (comparison benchmark is in the works, but it's fast for the same reason that MixedModels.jl is faster than {lme4} - this is [known](https://dm13450.github.io/2022/01/06/Mixed-Models-Benchmarking.html)). jlmerclusterperm also exposes the individual steps of the CPA as independent functions, emphasizing the modular nature of the procedure and making it easier to interpret and diagnose a CPA.

Not applicable - package does not store or transmit any data

The package integrates Julia via [JuliaConnectoR](https://github.com/stefan-m-lenz/JuliaConnectoR) - it both calls existing Julia libraries as well as custom Julia code in `inst/julia`. The pkgcheck is set up via GHA and tracked in https://github.com/yjunechoe/jlmerclusterperm/issues/1. The [package website](https://yjunechoe.github.io/jlmerclusterperm/) includes many vignettes/articles that are not included in the package proper
