+++ CITATION.cff
@@ -1,0 +1,419 @@
+# -----------------------------------------------------------
+# CITATION file created with {cffr} R package, v0.5.0
+# See also: https://docs.ropensci.org/cffr/
+# -----------------------------------------------------------
+ 
+cff-version: 1.2.0
+message: 'To cite package "jlmerclusterperm" in publications use:'
+type: software
+license: MIT
+title: 'jlmerclusterperm: Bootstrapped cluster-based permutation analysis for time
+  series data using GLM and MixedModels in Julia'
+version: 0.2.0
+abstract: A Julia implementation of fast cluster-based permutation analysis (CPA)
+  for densely-sampled time data developed in Maris & Oostenveld, 2007 <doi:10.1016/j.jneumeth.2007.03.024>.
+  Uses (generalized, mixed-effects) regression models for the calculation of timewise
+  statistics. Provides both a wholesale and a piecemeal interface to the CPA procedure
+  with an emphasis on interpretability and diagnostics. Integrates Julia libraries
+  MixedModels and GLM for performance improvements. Provides additional functionalities
+  for R-Julia interface, powered by the JuliaConnectoR package (requires Julia installation).
+authors:
+- family-names: Choe
+  given-names: June
+  email: jchoe001@gmail.com
+  orcid: https://orcid.org/0000-0002-0701-921X
+repository-code: https://github.com/yjunechoe/jlmerclusterperm
+url: https://yjunechoe.github.io/jlmerclusterperm/
+contact:
+- family-names: Choe
+  given-names: June
+  email: jchoe001@gmail.com
+  orcid: https://orcid.org/0000-0002-0701-921X
+references:
+- type: software
+  title: 'R: A Language and Environment for Statistical Computing'
+  notes: Depends
+  url: https://www.R-project.org/
+  authors:
+  - name: R Core Team
+  location:
+    name: Vienna, Austria
+  year: '2023'
+  institution:
+    name: R Foundation for Statistical Computing
+  version: '>= 3.5'
+- type: software
+  title: cli
+  abstract: 'cli: Helpers for Developing Command Line Interfaces'
+  notes: Imports
+  url: https://cli.r-lib.org
+  repository: https://CRAN.R-project.org/package=cli
+  authors:
+  - family-names: Csárdi
+    given-names: Gábor
+    email: csardi.gabor@gmail.com
+  year: '2023'
+- type: software
+  title: generics
+  abstract: 'generics: Common S3 Generics not Provided by Base R Methods Related to
+    Model Fitting'
+  notes: Imports
+  url: https://generics.r-lib.org
+  repository: https://CRAN.R-project.org/package=generics
+  authors:
+  - family-names: Wickham
+    given-names: Hadley
+    email: hadley@rstudio.com
+  - family-names: Kuhn
+    given-names: Max
+    email: max@rstudio.com
+  - family-names: Vaughan
+    given-names: Davis
+    email: davis@rstudio.com
+  year: '2023'
+- type: software
+  title: JuliaConnectoR
+  abstract: 'JuliaConnectoR: A Functionally Oriented Interface for Integrating ''Julia''
+    with R'
+  notes: Imports
+  repository: https://CRAN.R-project.org/package=JuliaConnectoR
+  authors:
+  - family-names: Lenz
+    given-names: Stefan
+    email: stefan-m-lenz@web.de
+    orcid: https://orcid.org/0000-0001-9135-1743
+  - family-names: Binder
+    given-names: Harald
+    orcid: https://orcid.org/0000-0002-5666-8662
[…]
