+++ R/compute_timewise_statistics.R
@@ -92,7 +92,9 @@ alert_diagnostics <- function(jlmer_spec, out) {
     if (any(singular_fits)) {
       cli::cli_alert_info("{.val {sum(singular_fits)}} singular fit{?s} ({round(mean(singular_fits) * 100, 2)}%).")
     }
-    re_n_terms <- sapply(lme4::findbars(jlmer_spec$formula$jl), function(x) setNames(length(x[[2]]), deparse1(x[[3]])))
+    re_n_terms <- vapply(lme4::findbars(jlmer_spec$formula$jl), function(x) {
+      setNames(length(x[[2]]), deparse1(x[[3]]))
+    }, character(1))
     if (mean(singular_fits) > .2 && any(re_n_terms > 1)) {
       cli::cli_alert_info("Average number of components estimated to capture 95% of RE variance:")
       rePCs <- rowMeans(out$rePCA_95_matrix)
