+++ DESCRIPTION
@@ -1,19 +1,18 @@
 Package: jlmerclusterperm
-Title: Bootstrapped cluster-based permutation analysis for time series
-    data using GLM and MixedModels in Julia
+Title: Cluster-based Permutation Analysis for Densely Sampled Time Data
 Version: 0.2.0
 Authors@R: 
     person("June", "Choe", , "jchoe001@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0002-0701-921X"))
-Description: A Julia implementation of fast cluster-based permutation
+Description: An implementation of fast cluster-based permutation
     analysis (CPA) for densely-sampled time data developed in Maris &
-    Oostenveld, 2007 <doi:10.1016/j.jneumeth.2007.03.024>.  Uses
+    Oostenveld, 2007 <doi:10.1016/j.jneumeth.2007.03.024>. Supports
     (generalized, mixed-effects) regression models for the calculation of
     timewise statistics. Provides both a wholesale and a piecemeal
     interface to the CPA procedure with an emphasis on interpretability
-    and diagnostics.  Integrates Julia libraries MixedModels and GLM for
-    performance improvements.  Provides additional functionalities for
-    R-Julia interface, powered by the JuliaConnectoR package (requires
+    and diagnostics. Integrates Julia libraries 'MixedModels' and 'GLM' for
+    performance improvements, with additional functionalities for
+    R-Julia interface powered by the 'JuliaConnectoR' package (requires
     Julia installation).
 License: MIT + file LICENSE
 URL: https://github.com/yjunechoe/jlmerclusterperm,
