+++ tests/testthat/test-singularity.R
@@ -1,6 +1,6 @@
 singularity_spec <- make_jlmer_spec(
   formula = weight ~ 1 + Diet + (1 + Diet | Chick),
-  data = subset(transform(chickweights, Diet = as.integer(Chick)), Chick %in% 1:5),
+  data = subset(transform(ChickWeight, Diet = as.integer(Chick)), Chick %in% 1:5 & Time <= 20),
   subject = "Chick", time = "Time"
 )
 
