+++ tests/testthat/test-jlmer_spec.R
@@ -7,3 +7,7 @@ test_that("Inform misspecifications in grouping structure", {
   expect_message(make_jlmer_spec(weight ~ 1 + Diet, subset(ChickWeight, Time <= 20), time = "Time"))
   expect_message(make_jlmer_spec(weight ~ 1 + Diet, ChickWeight, subject = "Chick", time = "Time"))
 })
+
+test_that("Warn uneven time sampling rate", {
+  expect_message(make_jlmer_spec(weight ~ 1 + Diet, ChickWeight, subject = "Chick", time = "Time"), "Sampling rate")
+})
