+++ man/jlmerclusterperm-package.Rd
@@ -4,11 +4,11 @@
 \name{jlmerclusterperm-package}
 \alias{jlmerclusterperm}
 \alias{jlmerclusterperm-package}
-\title{jlmerclusterperm: Bootstrapped cluster-based permutation analysis for time series data using GLM and MixedModels in Julia}
+\title{jlmerclusterperm: Cluster-based Permutation Analysis for Densely Sampled Time Data}
 \description{
 \if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
 
-A Julia implementation of fast cluster-based permutation analysis (CPA) for densely-sampled time data developed in Maris & Oostenveld, 2007 \doi{10.1016/j.jneumeth.2007.03.024}. Uses (generalized, mixed-effects) regression models for the calculation of timewise statistics. Provides both a wholesale and a piecemeal interface to the CPA procedure with an emphasis on interpretability and diagnostics. Integrates Julia libraries MixedModels and GLM for performance improvements. Provides additional functionalities for R-Julia interface, powered by the JuliaConnectoR package (requires Julia installation).
+An implementation of fast cluster-based permutation analysis (CPA) for densely-sampled time data developed in Maris & Oostenveld, 2007 \doi{10.1016/j.jneumeth.2007.03.024}. Supports (generalized, mixed-effects) regression models for the calculation of timewise statistics. Provides both a wholesale and a piecemeal interface to the CPA procedure with an emphasis on interpretability and diagnostics. Integrates Julia libraries 'MixedModels' and 'GLM' for performance improvements, with additional functionalities for R-Julia interface powered by the 'JuliaConnectoR' package (requires Julia installation).
 }
 \seealso{
 Useful links:
