+++ R/compute_timewise_statistics.R
@@ -14,7 +14,6 @@
 #'  These can be suppressed via the `suppress*()` functions.
 #'
 #' @examples
-#' \dontrun{
 #' library(dplyr, warn.conflicts = FALSE)
 #' jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 #'
@@ -40,7 +39,6 @@
 #' to_jlmer(weight ~ 1 + Diet, filter(ChickWeight, Time == 0)) %>%
 #'   tidy() %>%
 #'   select(term, statistic)
-#' }
 #'
 #' @return A predictor-by-time matrix of cluster statistics, of class `timewise_statistics`.
 #' @export
