### 2023-05-30 2c9f6e8
error logic for cluster comparison
