+++ man/jlmerclusterperm_setup.Rd
@@ -21,8 +21,6 @@ TRUE
 Initial setup for the jlmerclusterperm package
 }
 \examples{
-\dontrun{
 jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
-}
 
 }
