+++ man/clusterpermute.Rd
@@ -60,7 +60,6 @@ A list of \code{null_cluster_dists} and \code{empirical_clusters} with p-values
 Conduct a cluster-based permutation test
 }
 \examples{
-\dontrun{
 library(dplyr, warn.conflicts = FALSE)
 jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 
@@ -82,7 +81,6 @@ sapply(CPA, class)
 # You can extract the individual components for further inspection
 CPA$null_cluster_dists
 CPA$empirical_clusters
-}
 
 }
 \seealso{
