+++ R/calculate_pvalue.R
@@ -9,6 +9,7 @@
 #' @seealso [extract_empirical_clusters()], [extract_null_cluster_dists()]
 #'
 #' @examples
+#' \dontrun{
 #' library(dplyr, warn.conflicts = FALSE)
 #' jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 #'
@@ -47,6 +48,7 @@
 #' # Check for compatibility with `clusters_are_comparable()`
 #' clusters_are_comparable(empirical_clusters, null_cluster_dists)
 #' clusters_are_comparable(empirical_clusters2, null_cluster_dists)
+#' }
 #'
 #' @export
 #' @return An `empirical_clusters` object augmented with p-values.
