+++ R/interop-utils.R
@@ -18,6 +18,10 @@
 #' old_progress_opts
 #' julia_progress(old_progress_opts)
 #' identical(julia_progress(), old_progress_opts)
+#'
+#' # Alternatively, reset to default settings using this syntax:
+#' julia_progress(show = TRUE, width = "auto")
+#'
 #' }
 #'
 #' @return Previous values for `show` and `width`
