+++ vignettes/articles/Geller-et-al-2020.Rmd
@@ -21,7 +21,7 @@ The data comes from a lexical decision task using pupilometry (specifically, mea
 
 We will follow [Jason Geller's tutorial](https://www.drjasongeller.com/blog/2020-gp-pup/) which used the [clusterperm](https://github.com/dalejbarr/clusterperm/) package to conduct a cluster-based permutation analysis (CPA) of differences in pupil size between the print and cursive conditions.
 
-The data from the package (`cursive_agg_data`) looks slightly different from that used in the tutorial, so to ensure full reproducibility we use the [snapshot of the data](https://raw.githubusercontent.com/jgeller112/drjasongeller/main/content/blog/2020-07-10-CBPT/blog_data.csv) that the tutorial used:
+The data from the package (`cursive_agg_data`) looks slightly different from that used in the tutorial, so to ensure full reproducibility we use the [exact data](https://raw.githubusercontent.com/jgeller112/drjasongeller/main/content/blog/2020-07-10-CBPT/blog_data.csv) that the tutorial used:
 
 ```{r data, message = FALSE}
 library(dplyr)
@@ -38,7 +38,7 @@ The data comes prepared for this analysis out of the box. The following columns 
 
 As the name of the `cusrive_agg_data` variable suggests, the data has been aggregated within subject, collapsing across trials.
 
-The following reproduces the figure from the tutorial (it looks slightly different because [the data has been updated since](https://github.com/dmirman/gazer/commit/091d38f586640831dbfd87aae29965ecb0d2e253)):
+The following reproduces the figure from the tutorial:
 
 ```{r script_fig, message = FALSE}
 library(ggplot2)
@@ -47,7 +47,10 @@ script_fig <- ggplot(cursive_agg_data, aes(timebins, aggbaseline)) +
 script_fig
 ```
 
-Note that for within-participant predictors like `script`, the permutation algorithm will randomly swap the labels for condition between the trials from each subject. This preserves the temporal structure of the data (no swapping below the trial-level grouping) and subject-level groupings of the data (no swapping of trials across participants).
+<div class="callout" style="font-size:0.9em">
+For within-participant predictors like `script`, the permutation algorithm will randomly swap the labels for condition _between_ the trials _within_ each subject. This preserves the temporal structure of the trial-level data (no swapping below the trial-level grouping) as well as the subject-level grouping structure (no swapping of trials across participants).
+</div>
+
 
 ## Outline
 
@@ -216,12 +219,12 @@ timewise_fig
 We find the same clusters identified between 2000ms-2300ms for both "t" and "chisq", with the peaks more pronounced for "chisq". The differences between the two are inconsequential for this example, but may produce different results in other cases.
 
 <div class="callout" style="font-size:0.9em">
-The chi-squared statistic (which `jlmerclusterperm` computes via a likelihood ratio test) is often preferred for testing single parameters because it makes less assumptions and tend to be more robust ([glmmFAQ](https://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#model-diagnostics)). But because goodness-of-fit tests are at the level of a predictor in a model formula, "chisq" is less interpretable for multi-level predictors (k-1 > 1). To tease apart the contribution of the individual levels of a multi-level predictor, `statistic = "t"` may be more appropriate.
[…]
