### 2023-05-30 00a0fa2
remove platform dependent check
