+++ tests/testthat/test-jlmer.R
@@ -1,3 +1,5 @@
+skip_conditionally()
+
 #' @srrstats {G5.3} Tests for correctness but no explicit test for NA
 #' @srrstats {G5.4} Tests compare to R `lm()` output.
 
