+++ tests/testthat/test-julia_rng.R
@@ -1,3 +1,5 @@
+skip_conditionally()
+
 jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 
 test_that("RNG counter setter/getter", {
