+++ tests/testthat/test-helper-skip.R
@@ -1,0 +1,7 @@
+skip_conditionally <- function() {
+  skip_on_cran()
+  if (!JuliaConnectoR::juliaSetupOk()) {
+    skip("Julia setup not okay.")
+  }
+  invisible()
+}
