+++ tests/testthat/test-singularity.R
@@ -1,3 +1,7 @@
+skip_conditionally()
+
+jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
+
 singularity_spec <- make_jlmer_spec(
   formula = weight ~ 1 + Diet + (1 + Diet | Chick),
   data = subset(transform(ChickWeight, Diet = as.integer(Chick)), Chick %in% 1:5 & Time <= 20),
