+++ tests/testthat/helper-skip.R
@@ -1,6 +1,7 @@
+no_julia <- !JuliaConnectoR::juliaSetupOk()
 skip_conditionally <- function() {
   skip_on_cran()
-  if (!JuliaConnectoR::juliaSetupOk()) {
+  if (no_julia) {
     skip("Julia setup not okay.")
   }
   invisible()
