+++ tests/testthat/test-julia_rng.R
@@ -18,3 +18,11 @@ test_that("RNG restore", {
   expect_equal(get_rng_seed(), 1)
   expect_equal(get_rng_state(), 0)
 })
+
+test_that("RNG generate random seed", {
+  expect_message(rand_seed <- set_rng_seed(), "Using randomly generated seed")
+  expect_equal(rand_seed, get_rng_seed())
+})
+
+set_rng_seed(1)
+reset_rng_state()
