+++ R/utils.R
@@ -29,9 +29,7 @@ zero_pad <- function(x, y) {
 
 check_arg_class <- function(x, x_class, x_arg = x_class) {
   if (!inherits(x, x_class)) {
-    cli::cli_abort(c(
-      "{.arg {x_arg}} must be a {.cls {x_class}} object, not a {.cls {class(x)}}"
-    ))
+    cli::cli_abort("{.arg {x_arg}} must be a {.cls {x_class}} object, not a {.cls {class(x)}}")
   }
   invisible(TRUE)
 }
