+++ man/clusterpermute.Rd
@@ -60,7 +60,7 @@ A list of \code{null_cluster_dists} and \code{empirical_clusters} with p-values
 Conduct a cluster-based permutation test
 }
 \examples{
-\dontshow{if (JuliaConnectoR::juliaSetupOk()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontrun{
 library(dplyr, warn.conflicts = FALSE)
 jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 
@@ -82,11 +82,8 @@ sapply(CPA, class)
 # You can extract the individual components for further inspection
 CPA$null_cluster_dists
 CPA$empirical_clusters
-
-\dontshow{
-JuliaConnectoR::stopJulia()
 }
-\dontshow{\}) # examplesIf}
+
 }
 \seealso{
 \code{\link[=compute_timewise_statistics]{compute_timewise_statistics()}}, \code{\link[=permute_timewise_statistics]{permute_timewise_statistics()}},
