+++ DESCRIPTION
@@ -40,7 +40,6 @@ Suggests:
     knitr,
     MASS,
     patchwork,
-    peekbankr,
     readr,
     rmarkdown,
     scales,
@@ -48,8 +47,6 @@ Suggests:
     tibble
 VignetteBuilder: 
     knitr
-Remotes: 
-    langcog/peekbankr
 Config/testthat/edition: 3
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))

[…]
