+++ README.Rmd
@@ -48,7 +48,7 @@ library(jlmerclusterperm)
 system.time(jlmerclusterperm_setup(verbose = FALSE))
 ```
 
-See the [Articles](https://yjunechoe.github.io/jlmerclusterperm/articles/) page on the [package website](https://yjunechoe.github.io/jlmerclusterperm/) for tutorials and case studies.
+See the [Get Started](https://yjunechoe.github.io/jlmerclusterperm/articles/jlmerclusterperm.html) page on the [package website](https://yjunechoe.github.io/jlmerclusterperm/) for background and tutorials.
 
 ## Quick tour of package functionalities
 
