### 2023-06-09 e6ca9e2
use 2 threads for tests
