+++ R/clusterpermute.R
@@ -9,10 +9,14 @@
 #'  [extract_empirical_clusters()], [extract_null_cluster_dists()],
 #'  [calculate_clusters_pvalues()]
 #'
-#' @examples
-#' \dontrun{
+#' @examplesIf JuliaConnectoR::juliaSetupOk()
+#' \dontshow{
+#' options("jlmerclusterperm.nthreads" = 2)
+#' jlmerclusterperm_setup(verbose = FALSE)
+#' julia_progress(show = FALSE)
+#' }
+#' \donttest{
 #' library(dplyr, warn.conflicts = FALSE)
-#' jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 #'
 #' # Specification object
 #' spec <- make_jlmer_spec(
@@ -32,6 +36,9 @@
 #' # You can extract the individual components for further inspection
 #' CPA$null_cluster_dists
 #' CPA$empirical_clusters
+#' }
+#' \dontshow{
+#' JuliaConnectoR::stopJulia()
 #' }
 #'
 #' @export
