+++ tests/testthat/test-singularity.R
@@ -1,6 +1,6 @@
-skip_conditionally()
+# skip_conditionally()
 
-jlmerclusterperm_setup(verbose = FALSE)
+jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 
 singularity_spec <- make_jlmer_spec(
   formula = weight ~ 1 + Diet + (1 + Diet | Chick),
