+++ tests/testthat/test-timewise_statistics.R
@@ -1,6 +1,6 @@
-skip_conditionally()
+# skip_conditionally()
 
-jlmerclusterperm_setup(verbose = FALSE)
+jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 
 spec <- make_jlmer_spec(
   weight ~ 1 + Diet, subset(ChickWeight, Time <= 20),
