+++ tests/testthat/test-threshold_search.R
@@ -1,6 +1,6 @@
-skip_conditionally()
+# skip_conditionally()
 
-jlmerclusterperm_setup(verbose = FALSE)
+jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 reset_rng_state()
 
 spec <- make_jlmer_spec(
