+++ man/julia_progress.Rd
@@ -18,6 +18,7 @@ Previous values for \code{show} and \code{width}
 Set/get options for Julia progress bar
 }
 \examples{
+\dontshow{if (JuliaConnectoR::juliaSetupOk()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 \dontshow{
 options("jlmerclusterperm.nthreads" = 2)
 jlmerclusterperm_setup(verbose = FALSE)
@@ -41,5 +42,5 @@ julia_progress(show = TRUE, width = "auto")
 \dontshow{
 JuliaConnectoR::stopJulia()
 }
-
+\dontshow{\}) # examplesIf}
 }
