+++ man/permute_by_predictor.Rd
@@ -27,6 +27,7 @@ A long dataframe of permuted re-samples with \code{.id} column representing repl
 Permute data while respecting grouping structure(s) of observations
 }
 \examples{
+\dontshow{if (JuliaConnectoR::juliaSetupOk()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 \dontshow{
 options("jlmerclusterperm.nthreads" = 2)
 jlmerclusterperm_setup(verbose = FALSE)
@@ -76,5 +77,5 @@ identical(spec2_perm1, spec2_perm2)
 \dontshow{
 JuliaConnectoR::stopJulia()
 }
-
+\dontshow{\}) # examplesIf}
 }
