+++ man/cluster_permutation_tidiers.Rd
@@ -25,6 +25,7 @@ A data frame
 Tidiers for cluster permutation test objects
 }
 \examples{
+\dontshow{if (JuliaConnectoR::juliaSetupOk()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 \dontshow{
 options("jlmerclusterperm.nthreads" = 2)
 jlmerclusterperm_setup(verbose = FALSE)
@@ -63,5 +64,5 @@ tidy(null_cluster_dists)
 \dontshow{
 JuliaConnectoR::stopJulia()
 }
-
+\dontshow{\}) # examplesIf}
 }
