+++ DESCRIPTION
@@ -49,8 +49,10 @@ VignetteBuilder:
     knitr
 Config/testthat/edition: 3
 Encoding: UTF-8
-Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
+Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
+    "srr::srr_stats_roclet"))
 RoxygenNote: 7.2.3
+SystemRequirements: Julia (>= 1.8)
 Collate:
     'jlmerclusterperm-package.R'
     'aaa.R'
