+++ tests/testthat/test-threshold_search.R
@@ -1,4 +1,4 @@
-# skip_conditionally()
+skip_conditionally()
 
 jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 reset_rng_state()
