+++ tests/testthat/helper-skip.R
@@ -1,6 +1,6 @@
 no_julia <- !JuliaConnectoR::juliaSetupOk()
 skip_conditionally <- function() {
-  skip_on_cran()
+  # skip_on_cran()
   if (no_julia) {
     skip("Julia setup not okay.")
   }
