+++ tests/testthat/test-julia_rng.R
@@ -1,4 +1,4 @@
-# skip_conditionally()
+skip_conditionally()
 
 jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)
 
