+++ vignettes/jlmerclusterperm.Rmd
@@ -26,7 +26,7 @@ CPA formalizes two intuitions about what it means for there to be a difference b
 
 2) The degree of extremity of a cluster (i.e., the **cluster-mass statistic**) is a measure that is sensitive to the magnitude of the difference, its variability, and the sample size (e.g., the t-statistic from a regression).
 
-The CPA procedure identifies empirical clusters in a time series data and tests the significance of the observed cluster-mass statistics against **bootstrapped permutations** of the data. The bootstrapping procedure samples from the null, yielding a distribution of cluster-mass statistics emerging from "chance" (i.e., random shuffling of condition labels). The statistical significance of a cluster is the probability of observing a cluster-mass statistic as extreme as the cluster's given the simulated null distribution. 
+The CPA procedure identifies empirical clusters in a time series data and tests the significance of the observed cluster-mass statistics against **bootstrapped permutations** of the data. The bootstrapping procedure samples from the "null" (via random shuffling of condition labels), yielding a distribution of cluster-mass statistics emerging from chance. The statistical significance of a cluster is the probability of observing a cluster-mass statistic as extreme as the cluster's against the simulated null distribution. 
 
 ## Package design
 
@@ -38,31 +38,31 @@ See the [function documentation](https://yjunechoe.github.io/jlmerclusterperm/re
 
 ## Organization of vignettes
 
-The package [vignettes](https://yjunechoe.github.io/jlmerclusterperm/articles/) are roughly divided into two groups: **topics** and **case studies**. If you are a researcher with data ready for analysis, it is recommended to go through the case studies first and pluck out the relevant bits for your own hypothesis.
+The package [vignettes](https://yjunechoe.github.io/jlmerclusterperm/articles/) are roughly divided into two groups: **topics** and **case studies**. If you are a researcher with data ready for analysis, it is recommended to go through the case studies first and pluck out the relevant bits for your own desired analysis
 
 In the order of increasing complexity, the **case studies** are:
 
-1) [Garrison et al. 2020](https://yjunechoe.github.io/jlmerclusterperm/articles/Garrison-et-al-2020.html), which introduces the package's interface to CPA and compares a wholesale vs. piecemeal approach to CPA. It also compares the use of linear vs. logistic regression for the calculation of timewise statistics.
+1) [Garrison et al. 2020](https://yjunechoe.github.io/jlmerclusterperm/articles/Garrison-et-al-2020.html), which introduces the package's functions for running a CPA, demonstrating both wholesale and piecemeal approachrs. It also compares the use of linear vs. logistic regression for the calculation of timewise statistics.
 
 2) [Geller et al. 2020](https://yjunechoe.github.io/jlmerclusterperm/articles/Geller-et-al-2020.html), which demonstrates the use of random effects and regression contrasts in the CPA specification. It also compares the use of t vs. chisq timewise statistics.
 
 3) [de Carvalho et al. 2021](https://yjunechoe.github.io/jlmerclusterperm/articles/deCarvalho-et-al-2021.html), which showcases using custom regression contrasts to test the relationship between multiple levels of a factor. It also discusses issues surrounding the complexity of the timewise regression model.
 
-The **topics** are not particular to one type of analysis and instead cover general package features:
+The **topics** cover general package features:
 
 - [Tidying output](https://yjunechoe.github.io/jlmerclusterperm/articles/tidying-output.html): `tidy()` and other ways of collecting `jlmerclusterperm` objects as tidy data.
 
[…]
