### 2023-06-13 #591 @noamross
Hello @yjunechoe and thanks for you inquiry! We took some time looking at this because packages wrapping Julia code are a new category for us, and the editorial board wanted to consider whether we could reasonably identify reviewers and provide a good review.  We've determined we would accept this for review, but it would be an experimental test case to see how it goes. As with other statistical packages, you would have to still show compliance with the general and regression categories of standards, and at least 50% of those standards have to be applicable.

We do have a question: A couple of editors did want to know why you opted to structure the package in this way, rather than put most of the code in a Julia package?  If you organized it that way, your methods would be available in Julia, and then you could have only a thin wrapper around that Julia package expose everything in R.  In any case, we think a clear statement of purpose/audience in the package documentation should include why it uses mixed languages in this way.
