### 2023-06-14 #591 @yjunechoe
Thank you @noamross and to all those involved for considering jlmerclusterperm! I understand that this would make a somewhat atypical submission, so I am especially thankful for your evaluation. I suspect that there will be more hybrid packages of this sort in the future, so I'm happy for my package to be a guinea pig for an experimental review process.

To your question: this is largely a historical accident. I envisioned the package to be mostly written in R but simply swap out lm/lme4 for MixedModels. So originally I wrote a small Julia script to wrap one or two MixedModels.jl functions and do a for loop, but then throughout development I got more ambitious and started moving more of the package internals from R to Julia (and this has paid off in speed). In sum, the Julia code isn't itself a package because I simply did not start out doing so (and at the time it didn't make sense to, but at this point maybe it's better to do so?).

I do still have some reservations about whether it's necessary wrap the Julia code into a package, though (and it'd be helpful to get some second opinions). One reason is that most of the Julia functions are internal functions that are tightly integrated with R functions of the same name. As they're written, they're not suitable for a Julia user to pick up and use in Julia - they're very dependent on the particular inputs provided to them by the R functions. This was an intentional design: it allows me to validate the user input early and only once in the R functions and return nicely formatted messages from R. I'm happy with this very limited scope of the Julia code as simply powering my R functions, in part because I believe that a typical Julia user could/would just "do it themself" (my Julia code is just a glorified for loop).

I'm sure that there are still merits to turning the Julia code into a package even for its limited scope of being jlmerclusterperm's backend. In an ideal world I would simply do this, but due to my limited knowledge of Julia package development, I've opted to pursue a middle ground approach where the Julia code is sourced in a specific [project environment](https://pkgdocs.julialang.org/v1/environments/). One step further would be to wrap the functions into a [module](https://docs.julialang.org/en/v1/manual/modules/), which, among other things, would allow namespacing within Julia. Anyways, these are all just my half baked thoughts - I'm happy to discuss further in the actual submission issue.

Lastly, a question of my own if I may: the package is also undergoing review on CRAN. Would it be okay to open the actual submission issue for jlmerclusterperm after it's published on CRAN? CRAN's scrutiny of this package has been demanding and I would like to get this resolved before doing anything else with this package.

Best,
June
