### 2023-06-14 #591 @mpadge
@yjunechoe Regarding documentation of standards compliance, this could also serve as an experimental test case. In particular, you'll likely need to document compliance of your `inst/julia` code. There is currently no direct way to do that, so it would be good to work with you to devise a satisfactory solution for how that might be done. When you get that far, please ping me via an issue in your package, and we can discuss further how to adapt our system to optimise your documentation procedure. Thanks.

And I'll also answer your last question: We generally advise CRAN submission after review here, but if you've already started with CRAN, then yes, feel free to resolve that first before proceeding with full submission here.
