### 2023-06-16 #591 @yjunechoe
@mpadge Thank you! And before I close the issue, just to clarify: should I tag you on a new issue in my repo regarding the statistical standards prior to creating the submission issue here (or would that be part of the review process itself?).
