### 2023-06-16 #591 @mpadge
Yes @yjunechoe, please tag me in your own repo, not here, and we can continue discussions there. And yes, we'll need to use that to work out a sensible way to document standards as part of you preparing for a full submission.
